+++ src/libraries/utils.js
@@ -50,6 +50,7 @@ import * as damageAttribution from '../utils/damage-attribution.js';
 import * as panelGeometry from '../utils/panel-geometry.js';
 import * as choiceDialog from '../utils/choice-dialog.js';
 import * as simplePanel from '../utils/simple-panel.js';
+import * as consumableTarget from '../utils/consumable-target.js';
 import * as dropLuck from '../utils/drop-luck.js';
 import * as complexFft from '../utils/complex-fft.js';
 import * as combatDropModel from '../utils/combat-drop-model.js';
@@ -121,6 +122,7 @@ toolashaRoot.Utils = {
     panelGeometry,
     choiceDialog,
     simplePanel,
+    consumableTarget,
     dropLuck,
     complexFft,
     combatDropModel,